+++ HISTORY.md
@@ -31,6 +31,8 @@ Our backwards-compatibility policy can be found [here](https://github.com/python
   ([#707](https://github.com/python-attrs/cattrs/issues/707) [#708](https://github.com/python-attrs/cattrs/pull/708))
 - Enum handling has been optimized by switching to hook factories, improving performance especially for plain enums.
   ([#705](https://github.com/python-attrs/cattrs/pull/705))
+- Fix `include_subclasses` when used with `configure_tagged_union` and classes using diamond inheritance.
+  ([#685](https://github.com/python-attrs/cattrs/issues/685) [#713](https://github.com/python-attrs/cattrs/pull/713))
 
 ## 25.3.0 (2025-10-07)